添加相互转化的xml的文件信息
