+++ src/main/java/io/mycat/config/loader/zkprocess/entry/Named.java
@@ -1,0 +1,14 @@
+package io.mycat.config.loader.zkprocess.entry;
+
+/**
+ * presentation a object have a filed name.
+ */
+public interface Named {
+    /**
+     * 获得属性的名称
+    * 方法描述
+    * @return
+    * @创建日期 2016年9月15日
+    */
+    String getName();
+}
