+++ src/main/java/io/mycat/config/loader/zkprocess/entry/Propertied.java
@@ -1,0 +1,8 @@
+package io.mycat.config.loader.zkprocess.entry;
+
+/**
+ * Created by lion on 12/8/15.
+ */
+public interface Propertied {
+    void addProperty(Property property);
+}
