+++ src/main/java/io/mycat/config/loader/zkprocess/entry/package-info.java
@@ -1,0 +1,8 @@
+@XmlSchema(
+    xmlns = @XmlNs(prefix = "mycat", namespaceURI = "http://io.mycat/"),
+    elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+
+package io.mycat.config.loader.zkprocess.entry;
+
+import javax.xml.bind.annotation.XmlNs;
+import javax.xml.bind.annotation.XmlSchema;
