+++ src/main/java/io/mycat/config/loader/zkprocess/entry/schema/datahost/ReadHost.java
@@ -1,0 +1,43 @@
+package io.mycat.config.loader.zkprocess.entry.schema.datahost;
+
+import java.util.List;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlTransient;
+import javax.xml.bind.annotation.XmlType;
+
+/**
+ * <readHost host="" url="" password="" user=""></readHost>
+* 源文件名：ReadHost.java
+* 文件版本：1.0.0
+* 创建作者：liujun
+* 创建日期：2016年9月15日
+* 修改作者：liujun
+* 修改日期：2016年9月15日
+* 文件描述：TODO
+* 版权所有：Copyright 2016 zjhz, Inc. All Rights Reserved.
+*/
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "readHost")
+public class ReadHost extends WriteHost {
+
+    @XmlAttribute
+    protected String weight;
+
+    public String getWeight() {
+        return weight;
+    }
+
+    public void setWeight(String weight) {
+        this.weight = weight;
+    }
+
+    @XmlTransient
+    @Override
+    public List<ReadHost> getReadHost() {
+        return super.getReadHost();
+    }
+
+}
