--- src/main/java/io/mycat/config/loader/zktoxml/entry/Propertied.java
@@ -1,8 +1,0 @@
-package io.mycat.config.loader.zktoxml.entry;
-
-/**
- * Created by lion on 12/8/15.
- */
-public interface Propertied {
-    void addProperty(Property property);
-}
