--- src/main/java/io/mycat/config/loader/zktoxml/entry/package-info.java
@@ -1,8 +1,0 @@
-@XmlSchema(
-    xmlns = @XmlNs(prefix = "mycat", namespaceURI = "http://io.mycat/"),
-    elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-
-package io.mycat.config.loader.zktoxml.entry;
-
-import javax.xml.bind.annotation.XmlNs;
-import javax.xml.bind.annotation.XmlSchema;
